Shweet. Wrapped C++ call to generate body from network string.
Added a few helper calls inside network and iesorBody in order to be able to pass JSON object or string object (the C++ non-v8 versions). Then, wrapped the function call inside of the worldWrapper object.

Right now, everything is passed around as strings that are stringified and parsed as needed. In the future, it should be the actual objects.

The next step is to the send the parsed network into the physics world. Following that there will be a function to do a long simulation (with some type of recording taking place during the long simulation). Then a small function to wrap the drawing information, and for short simulations as well.

One final function will need to be written to load a custom physics environment.

Then we'll have 5 function entrances to the iesor world.

Load 2D environment
Load body from network
Simulate world (small steps)
Simulate world (long sim)
Draw current world.

That's pretty much everything we need out of the simulation. Once it's async, we'll be able to ignore this part of the library and focus on WIN, and UI.