Added new class to wrap around the iesorWorld and make bridge calls between WIN and the simulation. Pass in a network string, and this class will add to iesorWorld and return morphology and what have you.
This is like our experiment manager, in some sense. IESoRWorld doesn't need to know about where or how bodies are entered into it. It just needs to read the JSON.

Everything compiles and tests okay. Not very thorough testing. Next step is simulation wrapping, then sending that info to a UI for real validation.