Moved win control files into main directory for node module loading issues. Now have a UI file, a UI websocket controller, wired up to an express app used by the main winControl object.
That means printing to screen is soon. Very soon.

After simple printing of a single domain, we'll have to integrate full grid objects from the previous WIN code. All old WIN code will need to be ported in (win-save and win-gen etc etc). Then the HTML will need to be loaded as well.