UI now properly displaying a genome to network to body object.
Can animate (currently frozen for testing purposes). A test was added to verify genome to body generation is accurate (this required modifying the wrapper).

Websockets set up to handle events along with callbacks (makes for easier coding on the client side).

Next step: Actually wire up WIN to this setup. It shouldn't be too dificult actually. The hardest part is the grid display code, which should be re-written.

More to come...